requirements.txt
asgiref==3.8.1
Django==4.2
joblib==1.4.2
mpmath==1.3.0
numpy==1.26.4
pandas==2.2.3
pillow==11.0.0
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.2.2
scipy==1.14.1
six==1.16.0
sqlparse==0.5.2
sympy==1.13.3
threadpoolctl==3.5.0
tzdata==2024.2
